requirements.txt
ajax==3.0.0
asgiref==3.8.1
decorator==5.2.1
Django==5.2.1
django-allauth==65.8.0
django-appconf==1.0.6
djangorestframework==3.16.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
inflection==0.5.1
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
